+++ catalog/utilities/media_props_gen/helpers/md.py
